+++ config_data_structure.py
@@ -22,7 +22,7 @@ PAGINATION_DEFAULT = 500
 
 # Add support for CORS
 X_DOMAINS = '*'
-X_HEADERS = ['Access-Control-Allow-Origin']
+X_HEADERS = ['Access-Control-Allow-Origin', 'If-Match', 'Authorization', 'User-Agent', 'Content-Type']
 
 # Allow $regex filtering. Default config blocks where and regex.
 MONGO_QUERY_BLACKLIST = ['$where']
@@ -324,6 +324,10 @@ statistics_schema = {
         'type': 'string',
         'minlength': 1,
     },
+    'status': {
+        'type': 'string',
+        'minlength': 1,
+    },
     'nodes_total': {
         'type': 'integer',
     },
@@ -495,6 +499,9 @@ statistics_schema = {
             'no_valid_owner': {
                 'type': 'boolean',
             },
+            'count': {
+                'type': 'integer',
+            },
         },
     },
     'bundles': {
@@ -611,6 +618,9 @@ statistics_schema = {
                         'type': 'integer',
                     },
                 },
+            },
+            'other': {
+                'type': 'string',
             },
         },
     },