+++ tasks.py
@@ -12,6 +12,7 @@ from celery import Celery
 from celery import group
 from celery.utils.log import get_task_logger
 from fabric.api import execute
+from datetime import datetime, timedelta
 from atlas.config import *
 from atlas import utilities
 from atlas import config_celery
@@ -27,6 +28,10 @@ logger = get_task_logger(__name__)
 # Create the Celery app object
 celery = Celery('tasks')
 celery.config_from_object(config_celery)
+
+
+class CeleryException(Exception):
+    pass
 
 
 @celery.task
@@ -170,17 +175,28 @@ def instance_provision(instance):
     # Set future instance status for settings file creation.
     instance['status'] = 'available'
 
-    provision_task = execute(fabfile.instance_provision, instance=instance)
+    try:
+        provision_task = execute(fabfile.instance_provision, instance=instance)
+        if isinstance(provision_task.get('host_string', None), BaseException):
+            raise provision_task.get('host_string')
+    except CeleryException as e:
+        logger.info('Site provision failed | Error Message | %s', e.message)
 
-    logger.debug(provision_task)
-    logger.debug(provision_task.values)
+    logger.debug('Site provision | Provision Fabric task | %s', provision_task)
+    logger.debug('Site provision | Provision Fabric task values | %s', provision_task.values)
 
-    install_task = execute(fabfile.instance_install, instance=instance)
+    try:
+        install_task = execute(fabfile.instance_install, instance=instance)
+        if isinstance(install_task.get('host_string', None), BaseException):
+            raise install_task.get('host_string')
+    except CeleryException as e:
+        logger.info('Site install failed | Error Message | %s', e.message)
 
-    logger.debug(install_task)
-    logger.debug(install_task.values)
+    logger.debug('Site provision | Install Fabric task | %s', install_task)
+    logger.debug('Site provision | Install Fabric task values | %s', install_task.values)
 
-    patch_payload = {'status': 'available', 'db_key': instance['db_key'], 'statistics': instance['statistics']}
+    patch_payload = {'status': 'available',
+                     'db_key': instance['db_key'], 'statistics': instance['statistics']}
     patch = utilities.patch_eve('instance', instance['_id'], patch_payload)
 
     profile = utilities.get_single_eve('code', instance['code']['profile'])
@@ -190,8 +206,9 @@ def instance_provision(instance):
     core_string = core['meta']['name'] + '-' + core['meta']['version']
 
     provision_time = time.time() - start_time
-    logger.info('Atlas operational statistic | Instance Provision - {0} - {1} | {2} '.format(core_string, profile_string, provision_time))
-    logger.debug('Instance has been provisioned\n{0}'.format(patch))
+    logger.info('Atlas operational statistic | Instance Provision | %s | %s | %s ',
+                core_string, profile_string, provision_time)
+    logger.debug('Instance provision | Patch | %s', patch)
 
     slack_title = '{0}/{1}'.format(base_urls[environment], instance['path'])
     slack_link = '{0}/{1}'.format(base_urls[environment], instance['path'])
@@ -323,18 +340,23 @@ def instance_update(instance, updates, original):
 @celery.task
 def instance_remove(instance):
     """
-    Remove site from the server.
+    Remove site from the server and delete Statistic item.
 
     :param instance: Item to be removed.
     :return:
     """
-    logger.debug('Instance remove\n{0}'.format(instance))
+    logger.debug('Instance remove | %s', instance)
     if instance['type'] == 'express':
         # execute(fabfile.instance_backup, instance=instance)
         # Check if stats object exists first.
-        if instance.get('statistics'):
-            utilities.delete_eve('statistics', instance['statistics'])
+        statistics_query = 'where={{"instance":"{0}"}}'.format(instance['_id'])
+        statistics = utilities.get_eve('statistics', statistics_query)
+        logger.debug('Statistics | %s', statistics)
+        if not statistics['_meta']['total'] == 0:
+            for statistic in statistics['_items']:
+                utilities.delete_eve('statistics', statistic['_id'])
         execute(fabfile.instance_remove, instance=instance)
+    logger.debug('Site remove | %s', site)
 
     if environment != 'local':
         execute(fabfile.update_f5)
@@ -609,25 +631,24 @@ def delete_stuck_pending_instances():
     """
     instance_query = 'where={"status":"pending"}'
     instances = utilities.get_eve('sites', instance_query)
+    logger.debug('Pending instances | %s', sites)
     # Loop through and remove sites that are more than 30 minutes old.
     if not instances['_meta']['total'] == 0:
         for instance in instances['_items']:
             # Parse date string into structured time.
             # See https://docs.python.org/2/library/datetime.html#strftime-and-strptime-behavior
             # for mask format.
-            date_created = time.strptime(instance['_created'], "%Y-%m-%d %H:%M:%S %Z")
-            # Get time now, Convert date_created to seconds from epoch and
-            # calculate the age of the instance.
-            seconds_since_creation = time.time() - time.mktime(date_created)
-            logger.debug('{0} is {1} seconds old. Created: {2} Current: {3}'.format(
-                instance['sid'],
-                seconds_since_creation,
-                time.mktime(date_created),
-                time.time())
-                        )
-            # 15 min * 60 sec = 900 seconds
-            if seconds_since_creation > 900:
-                utilities.delete_eve('sites', site['_id'])
+            date_created = datetime.strptime(site['_created'], "%Y-%m-%d %H:%M:%S %Z")
+            # Get datetime now and calculate the age of the site. Since our timestamp is in GMT, we
+            # need to use UTC.
+            time_since_creation = datetime.utcnow() - date_created
+            logger.debug('%s has timedelta of %s. Created: %s Current: %s',
+                         site['sid'],
+                         time_since_creation,
+                         date_created,
+                         datetime.utcnow())
+            if time_since_creation > timedelta(minutes=15):
+                utilities.delete_eve('instance', instance['_id'])
 
 
 @celery.task
@@ -660,8 +681,10 @@ def delete_statistics_without_active_instance():
         if not instances['_meta']['total'] == 0:
             for instance in instances['_items']:
                 instance_id_list.append(instance['_id'])
+                logger.debug('Instance list | %s', instance_id_list)
         for statistic in statistics['_items']:
             if statistic['site'] not in instance_id_list:
+                logger.debug('Statistic not in list | %s', statistic['_id'])
                 utilities.delete_eve('statistics', statistic['_id'])
 
 
@@ -692,3 +715,72 @@ def take_down_installed_old_instances():
                 # Patch the status to 'take_down'.
                 payload = {'status': 'take_down'}
                 utilities.patch_eve('instance', instance['_id'], payload)
+
+
+@celery.task
+def verify_statistics():
+    """
+    Get a list of statistics items that have not been updated in 36 hours and notify users.
+    """
+    time_ago = datetime.utcnow() - timedelta(hours=36)
+    statistics_query = 'where={{"_updated":{{"$lte":"{0}"}}}}'.format(
+        time_ago.strftime("%Y-%m-%d %H:%M:%S GMT"))
+    outdated_statistics = utilities.get_eve('statistics', statistics_query)
+    logger.debug('Old statistics time | %s', time_ago.strftime("%Y-%m-%d %H:%M:%S GMT"))
+    logger.debug('outdated_statistics items | %s', outdated_statistics)
+    statistic_id_list = []
+    if not outdated_statistics['_meta']['total'] == 0:
+        for outdated_statistic in outdated_statistics['_items']:
+            statistic_id_list.append(outdated_statistic['_id'])
+
+        logger.debug('statistic_id_list | %s', statistic_id_list)
+
+        instance_query = 'where={{"_id":{{"$in":{0}}}}}'.format(json.dumps(statistic_id_list))
+        logger.debug('Instance query | %s', instance_query)
+        instances = utilities.get_eve('instance', instance_query)
+        instances_id_list = []
+        if not instances['_meta']['total'] == 0:
+            for instance in instances['_items']:
+                instances_id_list.append(instance['_id'])
+
+        slack_fallback = '{0} statistics items have not been updated in 36 hours.'.format(
+            len(statistic_id_list))
+        slack_link = '{0}/statistics?{1}'.format(base_urls[environment], instance_query)
+        slack_payload = {
+            "text": 'Outdated Statistics',
+            "username": 'Atlas',
+            "attachments": [
+                {
+                    "fallback": slack_fallback,
+                    "color": 'danger',
+                    "title": 'Some statistics items have not been updated in 36 hours.',
+                    "fields": [
+                        {
+                            "title": "Count",
+                            "value": len(statistic_id_list),
+                            "short": True
+                        },
+                        {
+                            "title": "Environment",
+                            "value": environment,
+                            "short": True
+                        },
+                    ],
+                },
+                {
+                    "fallback": 'Instance list',
+                    # A lighter red.
+                    "color": '#ee9999',
+                    "fields": [
+                        {
+                            "title": "Instance list",
+                            "value": json.dumps(instances_id_list),
+                            "short": False,
+                            "title_link": slack_link
+                        }
+                    ]
+                }
+            ],
+        }
+
+        utilities.post_to_slack_payload(slack_payload)