+++ CHANGELOG.md
@@ -17,10 +17,40 @@ This release migrates 'sites' to 'instance' and re-adds 'site' as an optional wr
 
 Resolves:
 
-- &#3590 - Create resource for backups
-- &#35230 - Convert site to instance
-- &#35231 - Add site
-- &#35232 - Fix Siteimprove parameters
+- &#35;90 - Create resource for backups
+- &#35;230 - Convert site to instance
+- &#35;231 - Add site
+- &#35;232 - Fix Siteimprove parameters
+
+## v1.0.17
+
+Resolves:
+
+- &#35;266 - Allow More Headers In Requests
+- &#35;271 - Add user count to statistics
+- &#35;270 - Add status to statistics
+- &#35;273 - Clean up temp files
+- &#35;268 - Statistic objects for stuck 'pending' instances are not cleaning up
+- &#35;265 - Verify statistics objects are getting updated
+- &#35;251 - Stuck pending instances not being removed
+- &#35;269 - Capture all bundles as typed or other
+- &#35;282 - Provision often fails, but the instance ends up 'available'. Settings files or DB creation fails. Switched to Fabric templating. Wrapped fabric commands in exception catching.
+
+## v1.0.16
+
+Resolves:
+
+- Instances are being marked as 'available' without databases.
+
+## v1.0.15
+
+Resolves:
+
+- &#35;245 - Available instances are not replaced each night
+- &#35;262 - Automatically clean up stats items that have deleted sites.
+- &#35;259 - Need to accept variable_livechat_license_number
+- &#35;256 - Code clone need labels
+- &#35;251 - Math for age of a pending instance is wrong
 
 ## v1.0.14